requirements.txt
agno
fastapi
uvicorn
qdrant-client
mcp
psycopg[binary]
python-dotenv
pydantic
sqlalchemy
requests
bcrypt
httpx
neo4j
python-jose[cryptography]
passlib[bcrypt]
